Refactoring + add moto requirement to test requirements
