+++ requirements_test.txt
@@ -9,3 +9,4 @@ black==19.10b0
 nbstripout==0.3.7
 pylint==2.5.2
 pre-commit==2.3.0
+moto==1.3.14
